feat: migrate tests
